gcp/requirements.txt
tensorflow==2.13.1
Pillow
numpy
google-cloud-storage